Assests for MainCharacte/Zombie/Wasp
These are the basic animation for the Main Character, Zombie and Wasp. For the Main Character there is; Attack animation, Dash animation, Damage animation, and Idle animation. For the Wasp there is; Basic model, Attack animation, and Idle animation. For the Zombie there is; Basic model, Attack animation, and Walk animation.